Move sliding sync types
